chore: moved React components to own file
